requirements.txt
streamlit
google-generativeai
yfinance
Pillow
pandas